moving classes to their own directories
