+++ game/constructs/end/EndGameConditions.groovy
@@ -1,4 +1,4 @@
-package constructs
+package constructs.end
 
 /**
  * @author Lawrence Thatcher
