+++ game/constructs/pieces/PiecePlacementType.groovy
@@ -1,4 +1,4 @@
-package constructs
+package constructs.pieces
 
 /**
  * @author Lawrence Thatcher
